\docType{data}
\encoding{UTF-8}


\name{Vignette_Param}
\alias{IniParam}
\alias{ListIniParam}
\alias{list_opt}
\alias{mcmcDRAM}
\alias{optDE}
\alias{optMALS}
\alias{optPORT}
\alias{optPORT_}
\alias{optPSO}
\alias{startGR4J}




\title{Datasets needed to run param_optim or param_mcmc vignettes}




\description{Datasets needed to run param_optim or param_mcmc vignettes.}




